+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Added null check in template config lookups
 * Init repo before init-package
 * Reload composer config after init-package
+* Generate CI workflow after init-package
 
 ## v0.1.13 (2022-11-21)
 * Fixed local specialised phpstan script calls